refactor: update project build output and wirte Makefile
