+++ Place.sln
@@ -50,8 +50,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core.EF", "src\Common\Core.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core.Domain", "src\Common\Core.Domain\Core.Domain.csproj", "{47DDA0F5-7152-4435-A5C8-51CCDCD61D49}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core.Web", "src\Common\Core.Web\Core.Web.csproj", "{B8AE7707-4869-43AF-B91E-D642133BC944}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core.MediatR", "src\Common\Core.MediatR\Core.MediatR.csproj", "{5453898F-4A78-4AD6-89B5-BAB204FF67FF}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Modules", "Modules", "{4B354D87-FCA4-46F3-A64E-76376D3C68C2}"
@@ -80,7 +78,6 @@ Global
 		{EA1DEC1C-397D-48A8-B85D-DC8D6363E0F3} = {25654B5F-D4E9-4019-B307-8759C38E94A8}
 		{7C3BC55D-F1DC-4CF8-9376-7941D59228AD} = {25654B5F-D4E9-4019-B307-8759C38E94A8}
 		{47DDA0F5-7152-4435-A5C8-51CCDCD61D49} = {25654B5F-D4E9-4019-B307-8759C38E94A8}
-		{B8AE7707-4869-43AF-B91E-D642133BC944} = {25654B5F-D4E9-4019-B307-8759C38E94A8}
 		{5453898F-4A78-4AD6-89B5-BAB204FF67FF} = {25654B5F-D4E9-4019-B307-8759C38E94A8}
 		{5B48E2F5-D7CC-48EA-94FE-A2F132650C8F} = {4B354D87-FCA4-46F3-A64E-76376D3C68C2}
 		{C3F80B30-C2DE-41DC-933A-A24E04827504} = {4B354D87-FCA4-46F3-A64E-76376D3C68C2}
@@ -141,10 +138,6 @@ Global
 		{47DDA0F5-7152-4435-A5C8-51CCDCD61D49}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{47DDA0F5-7152-4435-A5C8-51CCDCD61D49}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{47DDA0F5-7152-4435-A5C8-51CCDCD61D49}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B8AE7707-4869-43AF-B91E-D642133BC944}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B8AE7707-4869-43AF-B91E-D642133BC944}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B8AE7707-4869-43AF-B91E-D642133BC944}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B8AE7707-4869-43AF-B91E-D642133BC944}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5453898F-4A78-4AD6-89B5-BAB204FF67FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5453898F-4A78-4AD6-89B5-BAB204FF67FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5453898F-4A78-4AD6-89B5-BAB204FF67FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
