+++ docker/compose/databases/.env
@@ -1,6 +1,6 @@
 DB_USER=postgres
 DB_PASSWORD=postgres
-IDENTITY_DB=PlaceApiIdentity
-PROFILE_DB=PlaceApiProfile
+IDENTITY_DB=identity_db
+PROFILE_DB=profile_db
 DB_PORT=5489
 DB_HOST=localhost
