+++ src/Account/Profile/Features/V1/GetPersonalInformation/GetPersonalInformationEndpoint.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Net.Mime;
 using System.Threading.Tasks;
-using Core.Web;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
