+++ tests/Account.IntegrationTests/TestResults/integration_tests.trx
@@ -1,0 +1,27 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<TestRun id="e26c34dc-0295-438f-8518-8e54536536a0" name="@MacBook-Pro-de-Aristide 2024-11-23 22:04:28" xmlns="http://microsoft.com/schemas/VisualStudio/TeamTest/2010">
+  <Times creation="2024-11-23T22:04:28.0076740+01:00" queuing="2024-11-23T22:04:28.0076740+01:00" start="2024-11-23T22:04:27.4089730+01:00" finish="2024-11-23T22:04:28.0080020+01:00" />
+  <TestSettings name="default" id="ee958f0e-2053-4beb-8aa6-ce9045649df4">
+    <Deployment runDeploymentRoot="_MacBook-Pro-de-Aristide_2024-11-23_22_04_28" />
+  </TestSettings>
+  <TestLists>
+    <TestList name="Les résultats ne figurent pas dans une liste" id="8c84fa94-04c1-424b-9868-57a2d4851a1d" />
+    <TestList name="Tous les résultats chargés" id="19431567-8539-422a-85d7-44ee4e166bda" />
+  </TestLists>
+  <ResultSummary outcome="Completed">
+    <Counters total="0" executed="0" passed="0" failed="0" error="0" timeout="0" aborted="0" inconclusive="0" passedButRunAborted="0" notRunnable="0" notExecuted="0" disconnected="0" warning="0" completed="0" inProgress="0" pending="0" />
+    <Output>
+      <StdOut>[xUnit.net 00:00:00.00] xUnit.net VSTest Adapter v2.5.3.1+6b60a9e56a (64-bit .NET 9.0.0)
+[xUnit.net 00:00:00.06]   Discovering: Account.IntegrationTests
+[xUnit.net 00:00:00.09]   Discovered:  Account.IntegrationTests
+[xUnit.net 00:00:00.10]   Starting:    Account.IntegrationTests
+[xUnit.net 00:00:00.11]   Finished:    Account.IntegrationTests
+</StdOut>
+    </Output>
+    <RunInfos>
+      <RunInfo computerName="MacBook-Pro-de-Aristide" outcome="Warning" timestamp="2024-11-23T22:04:27.9448380+01:00">
+        <Text>Aucun test ne correspond au filtre testcase donné `Category=Integration` dans /Users/aristidehervembassi/Documents/Projects/Veepee/place-api/tests/Account.IntegrationTests/bin/Release/.net9.0/Account.IntegrationTests.dll</Text>
+      </RunInfo>
+    </RunInfos>
+  </ResultSummary>
+</TestRun>
