+++ tests/Identity.IntegrationTests/TestResults/unit_tests.trx
@@ -1,0 +1,27 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<TestRun id="5c93bf96-6c2a-46a5-8ad8-53230f00c8f7" name="@MacBook-Pro-de-Aristide 2024-11-23 22:09:01" xmlns="http://microsoft.com/schemas/VisualStudio/TeamTest/2010">
+  <Times creation="2024-11-23T22:09:01.0234840+01:00" queuing="2024-11-23T22:09:01.0234840+01:00" start="2024-11-23T22:09:00.6150980+01:00" finish="2024-11-23T22:09:01.0237280+01:00" />
+  <TestSettings name="default" id="9af54e2a-2fdb-4888-951a-d7cf39465610">
+    <Deployment runDeploymentRoot="_MacBook-Pro-de-Aristide_2024-11-23_22_09_01" />
+  </TestSettings>
+  <TestLists>
+    <TestList name="Les résultats ne figurent pas dans une liste" id="8c84fa94-04c1-424b-9868-57a2d4851a1d" />
+    <TestList name="Tous les résultats chargés" id="19431567-8539-422a-85d7-44ee4e166bda" />
+  </TestLists>
+  <ResultSummary outcome="Completed">
+    <Counters total="0" executed="0" passed="0" failed="0" error="0" timeout="0" aborted="0" inconclusive="0" passedButRunAborted="0" notRunnable="0" notExecuted="0" disconnected="0" warning="0" completed="0" inProgress="0" pending="0" />
+    <Output>
+      <StdOut>[xUnit.net 00:00:00.00] xUnit.net VSTest Adapter v2.5.3.1+6b60a9e56a (64-bit .NET 9.0.0)
+[xUnit.net 00:00:00.05]   Discovering: Identity.IntegrationTests
+[xUnit.net 00:00:00.07]   Discovered:  Identity.IntegrationTests
+[xUnit.net 00:00:00.07]   Starting:    Identity.IntegrationTests
+[xUnit.net 00:00:00.08]   Finished:    Identity.IntegrationTests
+</StdOut>
+    </Output>
+    <RunInfos>
+      <RunInfo computerName="MacBook-Pro-de-Aristide" outcome="Warning" timestamp="2024-11-23T22:09:00.9632460+01:00">
+        <Text>Aucun test ne correspond au filtre testcase donné `Category=Unit` dans /Users/aristidehervembassi/Documents/Projects/Veepee/place-api/tests/Identity.IntegrationTests/bin/Release/.net9.0/Identity.IntegrationTests.dll</Text>
+      </RunInfo>
+    </RunInfos>
+  </ResultSummary>
+</TestRun>
